Update licenses.md
Making some copy, clarity, and organizational changes to the license page.

Main deal: clarifying what's covered in the CC license (because we shouldn't claim to release non-original content we might reference, and not releasing images or code that isn't ours by just not releasing them).

The attribution line should be about "using material," without asing the user to think about what is the "correct" term for what sort of usage they're doing. Whether they're "extending" or "modifying" or just "pulling a couple paragraphs," "uses material from" covers all of that and more.